src/requirements.txt
flask
scrapy
psycopg2-binary
pandas
python-csv
